refactor: renamed files C2B => c2b
