+++ tests/unit/c2b/test_c2b.py
@@ -6,7 +6,7 @@ import warnings
 
 from mpesakit.auth import TokenManager
 from mpesakit.http_client import HttpClient
-from mpesakit.C2B import (
+from mpesakit.c2b import (
     C2B,
     C2BRegisterUrlRequest,
     C2BRegisterUrlResponse,
